# openapi/esim_go_schema_v2_5.yaml
openapi: 3.0.1
info:
  title: eSIM-Go API
  version: 2.5.0
  description: |
    To access the eSIMGo API, you need to authenticate your requests using an API key.
    Here's a brief overview of the authentication process:
      1. Assuming that you have created account already:
      - Log into your eSIMGo account at https://sso.esim-go.com/login.
      - Navigate to Account Settings -> API Details to find your API key.
      - Keep your API key secure and avoid sharing it with others.
      2. API Key Usage:
      - Include your API key in the header of all eSIMGo API requests.
      - Use the header key 'X-API-KEY' with your API key as the value.
      3. Security Scheme:
      - The eSIMGo API uses an HTTP security scheme type for authentication.
      4. Authorization Types:
      - The API supports 'apiKeyAuth' authorization type.
servers:
  - url: https://api.esim-go.com/v2.5
    description: eSIMGo API
...
paths:
...
  /organisation:
    get:
      summary: Get Current Organisation Details
      description: This endpoint allows you to retrieve comprehensive information about your organisation, including details of all associated users.
      tags:
        - Organisation
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationResponse"
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl 'https://api.esim-go.com/v2.5/organisation' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/organisation',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/organisation"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/organisation"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/organisation");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/organisation")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
  /organisation/balance:
    post:
      summary: Topup Organisation Balance
      description: |
        Update the pre-payed balance held by your organisation. Initial payment is done using the Portal: https://portal.esim-go.com/. This will store your card details for future use. Follow up payments can be done using the API.

         Note: The minimum top-up amount is $1000 and the maximum daily top-up is set at $5000. If you wish to raise or lower this amount, please contact your account manager.
      tags:
        - Organisation
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: amount
          in: query
          schema:
            type: string
          description: (Required) The amount of to be charged to the saved card
          example: 1000
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupResponse"
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      requestBody:
        content: {}
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl -X POST 'https://api.esim-go.com/v2.5/organisation/balance' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');
            let config = {
              method: 'post',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/organisation/balance',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/organisation/balance"
              method := "POST"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/organisation/balance"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("POST", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Post, "https://api.esim-go.com/v2.5/organisation/balance");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.post("https://api.esim-go.com/v2.5/organisation/balance")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
...
  /organisation/groups:
    get:
      summary: Get Bundle Groups
      description: |
        This endpoint returns a list of the groups of Bundles available for your Organisation to order. Bundles in eSIM Go are categorised into groups.

         Name of the bundle group depends on your account tier. In the example below you will find the groups description for **standard** organisation tier.

         Standard Fixed
         * All these bundles are for a finite and fixed amount of data.
         * eg. the moniker for a Standard Fixed Turkey 1GB 7 Days bundle is esim_1GB_7D_TR_V2

         Standard Long Duration
         * All these bundles are for a finite and fixed amount of data.
         * eg. the moniker for a Standard 50GB 90 Days Long Duration Bundle is esim_50GB_90D_RGBS_V2

         Standard Unlimited Lite
         * These bundles will provide 1GB of unthrottled data every 24 hours, once this is depleted data will be throttled to 512kbps. This repeats each day for the duration of the bundle.
         * eg. the moniker for a Standard Unlimited Lite Turkey 1 day bundle is esim_UL_1D_TR_V2

         Standard Unlimited Essential
         * These bundles will provide 1GB of unthrottled data every 24 hours, once this is depleted data will be throttled to 1.25mbps. This repeats each day for the duration of the bundle.
         * eg. the moniker for a Standard Unlimited Lite Turkey 1 day bundle is esim_ULE_1D_TR_V2

         Standard Unlimited Plus
         * These bundles will provide 2GB of unthrottled data every 24 hours, once this is depleted data will be throttled to 2mbps. This repeats each day for the duration of the bundle.
         * eg. the moniker for a Standard Unlimited Plus Turkey 1 day bundle is esim_ULP_1D_TR_V2
      tags:
        - Organisation
      security:
        - apiKeyAuth: []
      responses:
        "200":
          description: OK
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleGroupList"
              example:
                groups:
                  - name: Standard Fixed Bundles
                  - name: Standard Unlimited Bundles
        "400":
          description: Bad Request - returned when request format is not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Unauthorised - returned when the api token is invalid; the user does not have any available Bundles left or the ICCID is not accessible by the user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "503":
          description: Processing - Please come back later or use the *Retry-After* (seconds) header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
      x-codeSamples:
        - lang: shell
          label: cURL
          source: |-
            curl 'https://api.esim-go.com/v2.5/organisation/groups' \
            -H 'X-API-Key: $API_KEY' \
            -H 'Content-Type: application/json'
        - lang: javascript
          label: Node
          source: |
            const axios = require('axios');

            let config = {
              method: 'get',
              maxBodyLength: Infinity,
              url: 'https://api.esim-go.com/v2.5/organisation/groups',
              headers: { 
                'X-API-Key': '$API_KEY', 
                'Content-Type': 'application/json'
              }
            };

            axios.request(config)
            .then((response) => {
              console.log(JSON.stringify(response.data));
            })
            .catch((error) => {
              console.log(error);
            });
        - lang: go
          label: Go
          source: |-
            package main

            import (
              "fmt"
              "net/http"
              "io"
            )

            func main() {

              url := "https://api.esim-go.com/v2.5/organisation/groups"
              method := "GET"

              client := &http.Client {
              }
              req, err := http.NewRequest(method, url, nil)

              if err != nil {
                fmt.Println(err)
                return
              }
              req.Header.Add("X-API-Key", "$API_KEY")
              req.Header.Add("Content-Type", "application/json")

              res, err := client.Do(req)
              if err != nil {
                fmt.Println(err)
                return
              }
              defer res.Body.Close()

              body, err := io.ReadAll(res.Body)
              if err != nil {
                fmt.Println(err)
                return
              }
              fmt.Println(string(body))
            }
        - lang: python
          label: Python
          source: |
            import requests
            import json

            url = "https://api.esim-go.com/v2.5/organisation/groups"

            payload = {}
            headers = {
              'X-API-Key': '$API_KEY',
              'Content-Type': 'application/json'
            }

            response = requests.request("GET", url, headers=headers, data=payload)

            print(response.text)
        - lang: csharp
          label: C#
          source: |
            var client = new HttpClient();
            var request = new HttpRequestMessage(HttpMethod.Get, "https://api.esim-go.com/v2.5/organisation/groups");
            request.Headers.Add("X-API-Key", "$API_KEY");
            var content = new StringContent(string.Empty);
            content.Headers.ContentType = new MediaTypeHeaderValue("application/json");
            request.Content = content;
            var response = await client.SendAsync(request);
            response.EnsureSuccessStatusCode();
            Console.WriteLine(await response.Content.ReadAsStringAsync());
        - lang: java
          label: Java
          source: |
            Unirest.setTimeouts(0, 0);
            HttpResponse<String> response = Unirest.get("https://api.esim-go.com/v2.5/organisation/groups")
              .header("X-API-Key", "$API_KEY")
              .header("Content-Type", "application/json")
              .asString();
...
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      x-displayName: API Key in Header
      description: |
        Your API Key can passed to the eSIMGo API by setting a Header with the key `X-API-Key` in your HTTP call.

        e.g.

        ```
        curl -H 'X-API-Key: $API_KEY' https://api.esim-go.com/v2.3/esims/$ICCID/bundles
        ```
  schemas:
...
    Message:
      type: object
      properties:
        message:
          description: status of api functions which do not return data
          type: string
...
    TopupResponse:
      type: object
      properties:
        amount:
          type: integer
          description: The amount of the topup
          example: 5000
        balance:
          type: integer
          description: The current balance after the topup
          example: 12500
        ref:
          type: integer
          description: A reference number for the topup transaction
          example: 987654321
    OrganisationResponse:
      type: object
      properties:
        name:
          type: string
          example: Acme Corp
          description: Name of organisation
        apiKey:
          type: string
          example: ak_1234567890abcdef
          description: API Key
        taxLiable:
          type: string
          example: "Yes"
          description: If the organisation is tax liable
        addr1:
          type: string
          example: 123 Main St
          description: Address line 1
        addr2:
          type: string
          example: Suite 456
          description: Address line 2
        city:
          type: string
          example: Metropolis
          description: City
        country:
          type: string
          example: USA
          description: Country
        postcode:
          type: string
          example: "12345"
          description: Postal Code
        callbackUrl:
          type: string
          example: https://api.acmecorp.com/callback
          description: Callback URL
        notes:
          type: string
          example: Premium customer
          description: Notes attached to organisation
        groups:
          type: array
          items:
            type: string
          example:
            - Enterprise
            - Tech
          description: Groups an organisation assigned to
        currency:
          type: string
          example: USD
          description: Selected currency
        balance:
          type: integer
          example: 10000
          description: Organisational balance + test credit
        testCredit:
          type: integer
          example: 500
          description: Organisational test credit
        testCreditExpiry:
          type: string
          format: date
          example: "2023-12-31"
          description: Organisational test credit expiry date
        businessType:
          type: string
          example: Corporation
          description: Business type
        website:
          type: string
          example: https://www.acmecorp.com
          description: Website
        productDescription:
          type: string
          example: Cloud computing services
          description: Product Description
        users:
          type: array
          items:
            type: object
            properties:
              firstName:
                type: string
                example: John
                description: User first name
              lastName:
                type: string
                example: Doe
                description: User last name
              role:
                type: string
                example: Admin
                description: User role
              emailAddress:
                type: string
                example: "{EMAIL}"
                description: Email address (login)
              phoneNumber:
                type: string
                example: +1 (555) 123-4567
                description: Phone number
              timeZone:
                type: string
                example: America/New_York
                description: User Time Zone
...
    BundleGroupList:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/BundleGroup"
    BundleGroup:
      type: object
      properties:
        name:
          type: string
          description: Bundle group name
        priceListUrl:
          type: string
          format: uri
          description: URL to download a pricing list for specific bundle group
        desc:
          type: string
          description: Bundle group description
        icon:
          type: string
          format: uri
          description: URL to download a bundle group icon
      required:
        - name
...